Обновлена система приемной комиссии: увеличены окна, изменена на колледж, добавлена инструкция
1. Увеличены размеры всех окон для лучшей видимости:
   - LoginForm: 500x600 → 700x700 с увеличенными элементами и шрифтами
   - RegisterForm: 500x650 → 700x800 с увеличенными элементами и шрифтами
   - AdminPanel и UserPanel: 1200x700 → 1400x850

2. Добавлен файл инструкции в корень проекта:
   - Скопирован Инструкция_пользователя.docx в корень репозитория
   - Обновлен путь в LoginForm для поиска инструкции в корне проекта

3. Изменена система с ВУЗа на колледж:
   - Обновлены специальности: Информационные системы, Обеспечение безопасности,
     Дошкольное образование, Начальные классы
   - Изменены баллы ЕГЭ на средний балл аттестата (от 4.2 до 4.6)
   - Обновлены типы данных MinScore и ExamScore с INTEGER на REAL в базе данных
   - Обновлены модели Specialty и Application (int → double)
   - Обновлены все формы: заменены "Баллы ЕГЭ" на "Средний балл аттестата"
   - Настроен ввод баллов аттестата с диапазоном 2.0-5.0 и двумя десятичными знаками

4. Подтверждение работы админ-панели с логином admin/admin

+++ AdmissionSystem/Models/Application.cs
@@ -16,7 +16,7 @@ namespace AdmissionSystem.Models
         public string Address { get; set; }
         public string Phone { get; set; }
         public string Email { get; set; }
-        public int ExamScore { get; set; }
+        public double ExamScore { get; set; }
         public string Status { get; set; } // "На рассмотрении", "Одобрено", "Отклонено"
         public DateTime SubmissionDate { get; set; }
         public string Notes { get; set; }
